Some small misc. fixes
Deleted the rtl stuff that referred only to Bootstrap stuff.

Used named variables for break points in responsive.less

The filepicker stuff can be done in a renderer (core is doing
this for Bootstrap 2.3 classes in MDL-38903) and delete the
stuff at the top of responsive.

Need to redo the remains of responsive.less to be mobile first,
so that media queries are used only at larger sizes.