use nav-justified for responsive tabs
I handcoded a responsive hack to make tabs switch between
normal tabs on wide screens and "stacked" tabs at small sizes.

I filed a bug with Bootstrap for them to do this in v3. They
closed that bug, but it turns out they added basically the same
thing as long as you use nav-justified.

The switch to justified is more finger friendly on tablets anyway,
even when not stacked.

Nav pills also looks quite cool, someone should add a setting
for the theme to use those instead, it's just a swap of the class
name nav-tabs to nav-pills.

Oh, and I deleted some RTL added for Bootstrap upstream (some related
to nav, some not) in anticipation of using CSSflip.